refactor: move controller functions to caso.service.ts
